@@ -1,6 +1,10 @@
 name: nmobile
 description: nMobile
 
+# The following line prevents the package from being accidentally published to
+# pub.dev using `pub publish`. This is preferred for private packages.
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
 # followed by an optional build number separated by a +.
@@ -11,84 +15,57 @@ description: nMobile
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.2+137
+version: 1.0.0+1
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  flutter_cupertino_localizations: ^1.0.1
-  flutter_spinkit: ^4.1.1+1
-  convert: ^2.1.1
-  crypto: ^2.1.3
-  intl: ^0.16.0
-  dio: ^3.0.8
-  uuid: ^2.0.4
-  font_awesome_flutter: ^8.5.0
-  flutter_svg: ^0.19.0
-  shared_preferences: ^0.5.7+1
-  flutter_secure_storage: 3.3.5
-  local_auth: ^0.6.3+3
-  get_it: ^3.1.0
+
+  bot_toast: ^3.0.5
   bs58check: ^1.0.1
-  orientation: ^1.1.0
-  sqflite_sqlcipher: 1.0.0+5
-  path: ^1.6.4
-  path_provider: ^1.6.7
-  file_picker: ^1.4.3+2
-  image_picker: ^0.6.6+4
-  package_info: ^0.4.0+13
-  bloc: ^3.0.0
-  flutter_bloc: ^3.2.0
-  bot_toast: ^2.2.1
-  equatable: ^1.1.0
-  qr_flutter: ^3.2.0
-  qr_code_scanner: 0.0.12
-  share: ^0.6.3+6
+  dio: ^3.0.10
+  equatable: ^1.2.5
   esys_flutter_share: ^1.0.2
-  flutter_local_notifications: ^1.2.1
-  badges: ^1.1.1
-  flutter_app_badger: ^1.1.2
-  popup_menu: ^1.0.5
-  url_launcher: ^5.4.2
-  markdown: ^2.1.3
-  flutter_markdown: ^0.3.4
-  webview_flutter: ^0.3.22+1
-  mime_type: ^0.2.5
-  image_cropper: ^1.2.1
-  flutter_native_image: ^0.0.5+2
-  flutter_sound: ^2.1.1
-  background_fetch: ^0.5.6
-  common_utils: ^1.1.1
-  flustars: ^0.2.5+1
-  flutter_screenutil: ^1.1.0
-  flutter_swiper: ^1.1.6
-  cached_network_image: ^2.0.0-rc
-  oktoast: ^2.3.2
-  flutter_easyrefresh: ^2.1.1
-  flutter_easyloading: ^1.1.4
-  flutter_slidable: ^0.5.4
-  event_bus: ^1.1.0
-  flutter_image_compress: ^0.6.8
-  image_save: ^3.0.0
-  flutter_vibrate: ^1.0.0
-  # Logger
-  logger: ^0.9.1
-  sentry: ^3.0.1
+  file_picker: ^2.1.0
+  flutter_bloc: ^6.1.1
+  flutter_easyloading: ^2.2.1
+  flutter_local_notifications: ^3.0.2
+  flutter_secure_storage: ^3.3.5
+  flutter_svg: ^0.19.1
+  font_awesome_flutter: ^8.10.0
+  get_it: ^5.0.1
+  intl: ^0.16.1
+  image_picker: ^0.6.7+14
+  local_auth: ^0.6.3+4
+  package_info: ^0.4.3+2
+  path: ^1.7.0
+  path_provider: ^1.6.24
+  qr_code_scanner: 0.0.13
+  qr_flutter: ^3.2.0
+  shared_preferences: ^0.5.12+4
+  sqflite_sqlcipher: ^1.1.2
+  url_launcher: ^5.7.10
+  web3dart: ^1.2.3
+
+  # log
+  logger: ^0.9.4
+
+  # plugins
+  nkn_sdk_flutter:
+    path: ./plugins/nkn-sdk-flutter
+
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  web3dart: ^1.2.3
+  cupertino_icons: ^1.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  intl_translation: ^0.17.9
-
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -103,8 +80,8 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   # assets:
-  #  - images/a_dot_burr.jpeg
-  #  - images/a_dot_ham.jpeg
+  #   - images/a_dot_burr.jpeg
+  #   - images/a_dot_ham.jpeg
   assets:
     - assets/
     - assets/contact/
@@ -138,3 +115,5 @@ flutter:
   #
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
+flutter_intl:
+  enabled: true
