@@ -11,18 +11,14 @@ class FilteredWalletsBloc extends Bloc<FilteredWalletsEvent, FilteredWalletsStat
   final WalletsBloc walletsBloc;
   StreamSubscription walletsSubscription;
 
-  FilteredWalletsBloc({@required this.walletsBloc}) {
+  FilteredWalletsBloc({@required this.walletsBloc})
+      : super(walletsBloc.state is WalletsLoaded ? FilteredWalletsLoaded((walletsBloc.state as WalletsLoaded).wallets, null) : FilteredWalletsLoading()) {
     walletsBloc.listen((state) {
       if (state is WalletsLoaded) {
         add(UpdateWallets(state.wallets));
       }
     });
   }
-
-  @override
-  FilteredWalletsState get initialState => walletsBloc.state is WalletsLoaded
-      ? FilteredWalletsLoaded((walletsBloc.state as WalletsLoaded).wallets, null)
-      : FilteredWalletsLoading();
 
   @override
   Stream<FilteredWalletsState> mapEventToState(FilteredWalletsEvent event) async* {
@@ -37,11 +33,7 @@ class FilteredWalletsBloc extends Bloc<FilteredWalletsEvent, FilteredWalletsStat
     if (walletsBloc.state is WalletsLoaded) {
       var wallets = (walletsBloc.state as WalletsLoaded).wallets;
       yield FilteredWalletsLoaded(
-        event.filter != null
-            ? [wallets.singleWhere(event.filter, orElse: () => null)]
-            : [
-                null /*orElse*/
-              ],
+        event.filter != null ? [wallets.singleWhere(event.filter, orElse: () => null)] : [null],
         event.filter,
       );
     }
