@@ -1,23 +1,17 @@
 import 'package:bloc/bloc.dart';
 import 'package:nmobile/blocs/wallet/wallets_event.dart';
 import 'package:nmobile/blocs/wallet/wallets_state.dart';
-import 'package:nmobile/helpers/local_storage.dart';
-import 'package:nmobile/helpers/secure_storage.dart';
-import 'package:nmobile/model/db/upgrade_rn_wallet.dart';
 import 'package:nmobile/schemas/wallet.dart';
-import 'package:nmobile/utils/nlog_util.dart';
+import 'package:nmobile/storages/wallet.dart';
 
 class WalletsBloc extends Bloc<WalletsEvent, WalletsState> {
-  @override
-  WalletsState get initialState => WalletsLoading();
+  WalletsBloc() : super(WalletsLoading());
 
-  final LocalStorage _localStorage = LocalStorage();
-  final SecureStorage _secureStorage = SecureStorage();
+  final WalletStorage _walletStorage = WalletStorage();
 
   @override
   Stream<WalletsState> mapEventToState(WalletsEvent event) async* {
     if (event is LoadWallets) {
-      await _ensureRnWalletUpgraded();
       yield* _mapLoadWalletsToState();
     } else if (event is AddWallet) {
       yield* _mapAddWalletToState(event);
@@ -28,46 +22,13 @@ class WalletsBloc extends Bloc<WalletsEvent, WalletsState> {
     } else if (event is ReLoadWallets) {
       yield* _mapReloadWalletToState();
     } else if (event is UpdateWalletBackedUp) {
-      await _setWalletBackedUp(event);
-      yield* _mapReloadWalletToState();
-    }
-  }
-
-  _ensureRnWalletUpgraded() async {
-    final upgraded = await _localStorage.get(LocalStorage.RN_WALLET_UPGRADED);
-    if (upgraded == null || !upgraded) {
-      final list = await UpgradeRnWallet.rnWalletList;
-      if (list != null) {
-        for (RnWalletData w in list) {
-          final nkn = w.isEth ? w.tokenBalance : w.balance;
-          final eth = w.isEth ? w.balance : '0';
-          await _addWallet(WalletSchema(address: w.address, type: w.isEth ? WalletSchema.ETH_WALLET : WalletSchema.NKN_WALLET, name: w.name, balance: double.parse(nkn), balanceEth: double.parse(eth)), w.keystore);
-        }
-      }
-      await _localStorage.set(LocalStorage.RN_WALLET_UPGRADED, true);
+      yield* _setWalletBackedUp(event);
     }
   }
 
   Stream<WalletsState> _mapLoadWalletsToState() async* {
-    var wallets = await _localStorage.getArray(LocalStorage.NKN_WALLET_KEY);
-    if (wallets != null) {
-      final list = wallets.map((x) {
-        var wallet = WalletSchema(address: x['address'], type: x['type'], name: x['name']);
-        if (x['balance'] != null) {
-          wallet.balance = x['balance'] ?? 0;
-        }
-        if (x['balanceEth'] != null) {
-          wallet.balanceEth = x['balanceEth'] ?? 0;
-        }
-        if (x['isBackedUp'] != null) {
-          wallet.isBackedUp = x['isBackedUp'];
-        }
-        return wallet;
-      }).toList();
-      yield WalletsLoaded(list);
-    } else {
-      yield WalletsLoaded();
-    }
+    var wallets = await _walletStorage.getAllWallets();
+    yield WalletsLoaded(wallets);
   }
 
   Stream<WalletsState> _mapReloadWalletToState() async* {
@@ -80,9 +41,16 @@ class WalletsBloc extends Bloc<WalletsEvent, WalletsState> {
 
   Stream<WalletsState> _mapAddWalletToState(AddWallet event) async* {
     if (state is WalletsLoaded) {
-      final List<WalletSchema> list = List.from((state as WalletsLoaded).wallets)..add(event.wallet);
+      yield WalletsLoading();
+      final List<WalletSchema> list = List.from((state as WalletsLoaded).wallets);
+      int index = list.indexWhere((w) => w.address == event.wallet.address);
+      if (index > -1) {
+        list[index] = event.wallet;
+      } else {
+        list.add(event.wallet);
+      }
       yield WalletsLoaded(list);
-      _addWallet(event.wallet, event.keystore);
+      _walletStorage.addWallet(event.wallet, event.keystore);
     }
   }
 
@@ -92,72 +60,32 @@ class WalletsBloc extends Bloc<WalletsEvent, WalletsState> {
       int index = list.indexOf(event.wallet);
       if (index >= 0) {
         list[index] = event.wallet;
-        _setWallet(index, event.wallet);
+        _walletStorage.setWallet(index, event.wallet);
       }
-      yield WalletsLoaded(list);
+      yield* _mapReloadWalletToState();
     }
   }
 
   Stream<WalletsState> _mapDeleteWalletToState(DeleteWallet event) async* {
     if (state is WalletsLoaded) {
+      yield WalletsLoading();
       final List<WalletSchema> list = List.from((state as WalletsLoaded).wallets);
       int index = list.indexOf(event.wallet);
       list.removeAt(index);
       yield WalletsLoaded(list);
-      _deleteWallet(index, event.wallet);
+      _walletStorage.deleteWallet(index, event.wallet);
     }
   }
 
-  _setWalletBackedUp(UpdateWalletBackedUp event) async {
+  Stream<WalletsState> _setWalletBackedUp(UpdateWalletBackedUp event) async* {
     final address = event.address;
     final List<WalletSchema> list = List.from((state as WalletsLoaded).wallets);
     final wallet = list.firstWhere((w) => w.address == address, orElse: () => null);
     if (wallet != null) {
-      NLog.d('wallet != null: $wallet');
       int index = list.indexOf(wallet);
       wallet.isBackedUp = true;
-      await _setWallet(index, wallet);
+      await _walletStorage.setWallet(index, wallet);
     }
-  }
-
-  Future _addWallet(WalletSchema wallet, String keystore) async {
-    List<Future> futures = <Future>[];
-
-    Map<String, dynamic> data = {'name': wallet.name, 'type': wallet.type, 'address': wallet.address, 'isBackedUp': wallet.isBackedUp};
-    if (wallet.balance != null) {
-      data['balance'] = wallet.balance ?? 0;
-    }
-    if (wallet.balanceEth != null) {
-      data['balanceEth'] = wallet.balanceEth ?? 0;
-    }
-    var wallets = await _localStorage.getArray(LocalStorage.NKN_WALLET_KEY);
-    int index = wallets?.indexWhere((x) => x['address'] == wallet.address) ?? -1;
-    if (index < 0) {
-      futures.add(_localStorage.addItem(LocalStorage.NKN_WALLET_KEY, data));
-    } else {
-      futures.add(_localStorage.setItem(LocalStorage.NKN_WALLET_KEY, index, data));
-    }
-    futures.add(_secureStorage.set('${SecureStorage.NKN_KEYSTORES_KEY}:${wallet.address}', keystore));
-    await Future.wait(futures);
-  }
-
-  Future _setWallet(int n, WalletSchema wallet) async {
-//    NLog.d('wallet.isBackedUp: ${wallet.isBackedUp}');
-    List<Future> futures = <Future>[];
-    Map<String, dynamic> data = {'name': wallet.name, 'type': wallet.type, 'address': wallet.address, 'isBackedUp': wallet.isBackedUp};
-    if (wallet.balance != null) {
-      data['balance'] = wallet.balance ?? 0;
-    }
-    if (wallet.balanceEth != null) {
-      data['balanceEth'] = wallet.balanceEth ?? 0;
-    }
-    futures.add(_localStorage.setItem(LocalStorage.NKN_WALLET_KEY, n, data));
-    await Future.wait(futures);
-  }
-
-  Future _deleteWallet(int n, WalletSchema wallet) async {
-    List<Future> futures = <Future>[];
-    futures.add(_localStorage.removeItem(LocalStorage.NKN_WALLET_KEY, n));
-    await Future.wait(futures);
+    yield* _mapReloadWalletToState();
   }
 }
