@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
-import 'package:nmobile/consts/theme.dart';
+import 'package:nmobile/theme/theme.dart';
 
 class LoadingDialog extends StatefulWidget {
 
