@@ -4,15 +4,12 @@ import 'package:flutter_svg/svg.dart';
 import 'package:nmobile/blocs/wallet/filtered_wallets_bloc.dart';
 import 'package:nmobile/blocs/wallet/filtered_wallets_event.dart';
 import 'package:nmobile/blocs/wallet/filtered_wallets_state.dart';
-import 'package:nmobile/components/dialog/bottom.dart';
 import 'package:nmobile/components/label.dart';
-import 'package:nmobile/consts/colors.dart';
-import 'package:nmobile/consts/theme.dart';
+import 'package:nmobile/generated/l10n.dart';
 import 'package:nmobile/helpers/format.dart';
-import 'package:nmobile/l10n/localization_intl.dart';
 import 'package:nmobile/schemas/wallet.dart';
-import 'package:nmobile/utils/extensions.dart';
-import 'package:nmobile/utils/image_utils.dart';
+import 'package:nmobile/theme/theme.dart';
+import 'package:nmobile/utils/assets.dart';
 
 class WalletDropdown extends StatefulWidget {
   final WalletSchema schema;
@@ -42,10 +39,9 @@ class _WalletDropdownState extends State<WalletDropdown> {
 
   @override
   Widget build(BuildContext context) {
+    S _localizations = S.of(context);
     return InkWell(
       onTap: () {
-        // Disable selection, otherwise the page logic becomes very complicated.
-        // BottomDialog.of(context).showSelectWalletDialog(title: widget.title);
       },
       child: BlocBuilder<FilteredWalletsBloc, FilteredWalletsState>(
         builder: (context, state) {
@@ -61,18 +57,21 @@ class _WalletDropdownState extends State<WalletDropdown> {
                     flex: 0,
                     child: Stack(
                       children: [
-                        Container(
-                          width: 48,
-                          height: 48,
-                          alignment: Alignment.center,
-                          decoration: BoxDecoration(
-                            color: Colours.light_ff,
-                            borderRadius: BorderRadius.all(Radius.circular(8)),
+                        Padding(
+                          padding: const EdgeInsets.only(right: 16, top: 12, bottom: 12),
+                          child: Container(
+                            width: 48,
+                            height: 48,
+                            alignment: Alignment.center,
+                            decoration: BoxDecoration(
+                              color: DefaultTheme.lineColor,
+                              borderRadius: BorderRadius.all(Radius.circular(8)),
+                            ),
+                            child: SvgPicture.asset('assets/logo.svg', color: DefaultTheme.nknLogoColor),
                           ),
-                          child: SvgPicture.asset('assets/logo.svg', color: Colours.purple_2e),
-                        ).pad(r: 16, t: 12, b: 12),
+                        ),
                         wallet.type == WalletSchema.NKN_WALLET
-                            ? Space.empty
+                            ? Container()
                             : Positioned(
                                 top: 8,
                                 left: 32,
@@ -80,7 +79,7 @@ class _WalletDropdownState extends State<WalletDropdown> {
                                   width: 20,
                                   height: 20,
                                   alignment: Alignment.center,
-                                  decoration: BoxDecoration(color: Colours.purple_53, shape: BoxShape.circle),
+                                  decoration: BoxDecoration(color: DefaultTheme.ethLogoColor, shape: BoxShape.circle),
                                   child: SvgPicture.asset('assets/ethereum-logo.svg'),
                                 ),
                               )
@@ -93,7 +92,10 @@ class _WalletDropdownState extends State<WalletDropdown> {
                       mainAxisAlignment: MainAxisAlignment.center,
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: [
-                        Label(wallet.name, type: LabelType.h3).pad(b: 4),
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 4),
+                          child: Label(wallet.name, type: LabelType.h3),
+                        ),
                         Label(
                           Format.nknFormat(wallet.balance, decimalDigits: 4, symbol: 'NKN'),
                           type: LabelType.bodySmall,
@@ -110,51 +112,39 @@ class _WalletDropdownState extends State<WalletDropdown> {
                         wallet.type == WalletSchema.NKN_WALLET
                             ? Container(
                                 alignment: Alignment.center,
-                                padding: 2.pad(l: 8, r: 8),
-                                decoration: BoxDecoration(borderRadius: BorderRadius.all(Radius.circular(9)), color: Colours.green_06_a1p),
+                                padding: const EdgeInsets.only(top: 2, bottom: 2, left: 8, right: 8),
+                                decoration: BoxDecoration(borderRadius: BorderRadius.all(Radius.circular(9)), color: DefaultTheme.successColor.withAlpha(25)),
                                 child: Text(
-                                  NL10ns.of(context).mainnet,
-                                  style: TextStyle(color: Colours.green_06, fontSize: 10, fontWeight: FontWeight.bold),
+                                  _localizations.mainnet,
+                                  style: TextStyle(color: DefaultTheme.successColor, fontSize: 10, fontWeight: FontWeight.bold),
                                 ),
                               )
                             : Container(
                                 alignment: Alignment.center,
-                                padding: 2.pad(l: 8, r: 8),
-                                decoration: BoxDecoration(borderRadius: BorderRadius.all(Radius.circular(9)), color: Colours.purple_53_a1p),
+                                padding: const EdgeInsets.only(top: 2, bottom: 2, left: 8, right: 8),
+                                decoration: BoxDecoration(borderRadius: BorderRadius.all(Radius.circular(9)), color: DefaultTheme.ethLogoColor.withAlpha(25)),
                                 child: Text(
-                                  NL10ns.of(context).ERC_20,
-                                  style: TextStyle(color: Colours.purple_53, fontSize: 10, fontWeight: FontWeight.bold),
+                                  _localizations.ERC_20,
+                                  style: TextStyle(color: DefaultTheme.ethLogoColor, fontSize: 10, fontWeight: FontWeight.bold),
                                 ),
                               ),
-                        (wallet.type == WalletSchema.NKN_WALLET
-                                ? Space.empty
-                                : Label(
-                                    Format.nknFormat(wallet.balanceEth, symbol: 'ETH'),
-                                    type: LabelType.bodySmall,
-                                  ))
-                            .pad(t: 8),
-                      ],
-                    ),
-                  ),
-                  Expanded(
-                    flex: 0,
-                    child: Container(
-                      alignment: Alignment.centerRight,
-                      height: 44,
-                      child: Padding(
-                        padding: const EdgeInsets.only(left: 16),
-                        child: loadAssetIconsImage(
-                          'down2',
-                          width: 24,
+                        Padding(
+                          padding: const EdgeInsets.only(top: 8),
+                          child: (wallet.type == WalletSchema.NKN_WALLET
+                              ? Container()
+                              : Label(
+                                  Format.nknFormat(wallet.balanceEth, symbol: 'ETH'),
+                                  type: LabelType.bodySmall,
+                                )),
                         ),
-                      ),
+                      ],
                     ),
                   ),
                 ],
               ),
             );
           } else {
-            return null;
+            return Container();
           }
         },
       ),
