@@ -1,13 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:nmobile/components/label.dart';
-import 'package:nmobile/consts/colors.dart';
-import 'package:nmobile/consts/theme.dart';
+import 'package:nmobile/generated/l10n.dart';
 import 'package:nmobile/helpers/format.dart';
-import 'package:nmobile/l10n/localization_intl.dart';
 import 'package:nmobile/schemas/wallet.dart';
 import 'package:nmobile/screens/wallet/nkn_wallet_detail.dart';
-import 'package:nmobile/utils/extensions.dart';
+import 'package:nmobile/theme/theme.dart';
 
 class WalletItem extends StatefulWidget {
   final WalletSchema schema;
@@ -24,13 +22,13 @@ class WalletItem extends StatefulWidget {
 class _WalletItemState extends State<WalletItem> {
   @override
   Widget build(BuildContext context) {
+    S _localizations = S.of(context);
     if (widget.type == WalletType.nkn) {
       return InkWell(
         onTap: widget.onTap ??
             () {
-              Navigator.of(context).pushNamed(NknWalletDetailScreen.routeName, arguments: {
-                'wallet': widget.schema,
-                'index': widget.index,
+              Navigator.of(context).pushNamed(NknWalletDetailScreen.routeName,arguments: {
+                'wallet': widget.schema
               });
             },
         child: Container(
@@ -52,7 +50,7 @@ class _WalletItemState extends State<WalletItem> {
                       height: 48,
                       alignment: Alignment.center,
                       decoration: BoxDecoration(
-                        color: Color(0xFFF1F4FF),
+                        color: DefaultTheme.logoBackground,
                         borderRadius: BorderRadius.all(Radius.circular(8)),
                       ),
                       child: SvgPicture.asset('assets/logo.svg', color: Color(0xFF253A7E)),
@@ -99,7 +97,7 @@ class _WalletItemState extends State<WalletItem> {
                             alignment: Alignment.center,
                             padding: const EdgeInsets.only(left: 8, right: 8),
                             decoration: BoxDecoration(borderRadius: BorderRadius.all(Radius.circular(8)), color: Color(0x1500CC96)),
-                            child: Text(NL10ns.of(context).mainnet, style: TextStyle(color: Color(0xFF00CC96), fontSize: 10, fontWeight: FontWeight.bold)),
+                            child: Text(_localizations.mainnet, style: TextStyle(color: Color(0xFF00CC96), fontSize: 10, fontWeight: FontWeight.bold)),
                           ),
                         ],
                       ),
@@ -111,14 +109,12 @@ class _WalletItemState extends State<WalletItem> {
           ),
         ),
       );
-    } else {
-      assert(widget.type == WalletType.eth);
+    } else if (widget.type == WalletType.eth) {
       return InkWell(
         onTap: widget.onTap ??
             () {
               Navigator.of(context).pushNamed(NknWalletDetailScreen.routeName, arguments: {
                 'wallet': widget.schema,
-                'index': widget.index,
               });
             },
         child: Container(
@@ -126,34 +122,41 @@ class _WalletItemState extends State<WalletItem> {
             color: DefaultTheme.backgroundLightColor,
             borderRadius: BorderRadius.all(Radius.circular(12)),
           ),
-          child: Row(
+          child: Flex(
+            direction: Axis.horizontal,
             children: [
-              Hero(
-                tag: 'avatar:${widget.schema.address}',
-                child: Stack(
-                  children: [
-                    Container(
-                      width: 48,
-                      height: 48,
-                      alignment: Alignment.center,
-                      decoration: BoxDecoration(
-                        color: Colours.light_ff,
-                        borderRadius: BorderRadius.all(Radius.circular(8)),
+              Expanded(
+                flex: 0,
+                child: Hero(
+                  tag: 'avatar:${widget.schema.address}',
+                  child: Stack(
+                    children: [
+                      Container(
+                        padding: const EdgeInsets.symmetric(vertical: 20, horizontal: 16),
+                        child: Container(
+                          width: 48,
+                          height: 48,
+                          alignment: Alignment.center,
+                          decoration: BoxDecoration(
+                            color: DefaultTheme.logoBackground,
+                            borderRadius: BorderRadius.all(Radius.circular(8)),
+                          ),
+                          child: SvgPicture.asset('assets/logo.svg', color: DefaultTheme.nknLogoColor),
+                        ),
                       ),
-                      child: SvgPicture.asset('assets/logo.svg', color: Colours.purple_2e),
-                    ).symm(h: 16, v: 20),
-                    Positioned(
-                      top: 16,
-                      left: 48,
-                      child: Container(
-                        width: 20,
-                        height: 20,
-                        alignment: Alignment.center,
-                        decoration: BoxDecoration(color: Colours.purple_53, shape: BoxShape.circle),
-                        child: SvgPicture.asset('assets/ethereum-logo.svg'),
+                      Positioned(
+                        top: 12,
+                        left: 48,
+                        child: Container(
+                          width: 20,
+                          height: 20,
+                          alignment: Alignment.center,
+                          decoration: BoxDecoration(color: DefaultTheme.ethLogoColor, shape: BoxShape.circle),
+                          child: SvgPicture.asset('assets/ethereum-logo.svg'),
+                        ),
                       ),
-                    )
-                  ],
+                    ],
+                  ),
                 ),
               ),
               Expanded(
@@ -180,25 +183,28 @@ class _WalletItemState extends State<WalletItem> {
                 child: Container(
                   alignment: Alignment.centerRight,
                   height: 44,
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    crossAxisAlignment: CrossAxisAlignment.end,
-                    children: <Widget>[
-                      Container(
-                        alignment: Alignment.center,
-                        padding: 2.pad(l: 8, r: 8),
-                        decoration: BoxDecoration(borderRadius: BorderRadius.all(Radius.circular(9)), color: Colours.purple_53_a1p),
-                        child: Text(
-                          NL10ns.of(context).ERC_20,
-                          style: TextStyle(color: Colours.purple_53, fontSize: 10, fontWeight: FontWeight.bold),
+                  child: Padding(
+                    padding: const EdgeInsets.only(left: 12, right: 16),
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                      crossAxisAlignment: CrossAxisAlignment.end,
+                      children: <Widget>[
+                        Container(
+                          alignment: Alignment.center,
+                          padding: const EdgeInsets.only(left: 8, right: 8, top: 2, bottom: 2),
+                          decoration: BoxDecoration(borderRadius: BorderRadius.all(Radius.circular(9)), color: DefaultTheme.ethLogoColor.withAlpha(25)),
+                          child: Text(
+                            _localizations.ERC_20,
+                            style: TextStyle(color: DefaultTheme.ethLogoColor, fontSize: 10, fontWeight: FontWeight.bold),
+                          ),
                         ),
-                      ),
-                      Label(
-                        Format.nknFormat(widget.schema.balanceEth, symbol: 'ETH'),
-                        type: LabelType.bodySmall,
-                      ),
-                    ],
-                  ).pad(l: 12, r: 16),
+                        Label(
+                          Format.nknFormat(widget.schema.balanceEth, symbol: 'ETH'),
+                          type: LabelType.bodySmall,
+                        ),
+                      ],
+                    ),
+                  ),
                 ),
               ),
             ],
