@@ -1,101 +1,50 @@
-import 'dart:convert';
-
 import 'package:equatable/equatable.dart';
-import 'package:nmobile/components/dialog/bottom.dart';
-import 'package:nmobile/helpers/global.dart';
-import 'package:nmobile/helpers/local_notification.dart';
-import 'package:nmobile/helpers/local_storage.dart';
-import 'package:nmobile/helpers/secure_storage.dart';
-import 'package:nmobile/l10n/localization_intl.dart';
-import 'package:nmobile/plugins/nkn_wallet.dart';
-import 'package:nmobile/services/local_authentication_service.dart';
-import 'package:nmobile/utils/log_tag.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
 enum WalletType { nkn, eth }
 
-class WalletSchema extends Equatable with Tag {
+class WalletSchema extends Equatable {
   static const String NKN_WALLET = 'nkn';
   static const String ETH_WALLET = 'eth';
-  final String address;
-  final String type;
+  String address;
+  String type;
   String name;
-  double balance = 0;
+  double balance;
   String keystore;
+  double balanceEth;
+  bool isBackedUp;
 
-  //String publicKey; // fixme: null!!!
-  double balanceEth = 0;
-  bool isBackedUp = false;
+  WalletSchema({
+    this.address,
+    this.type,
+    this.name,
+    this.balance = 0,
+    this.balanceEth = 0,
+    this.isBackedUp = false,
+  });
 
-//  final LocalStorage _localStorage = LocalStorage();
-  final SecureStorage _secureStorage = SecureStorage();
+  WalletSchema.fromMap(Map<String, dynamic> map) {
+    this.address = map['address'];
+    this.type = map['type'];
+    this.name = map['name'] ?? '';
+    this.balance = map['balance'] ?? 0;
+    this.balanceEth = map['balanceEth'] ?? 0;
+    this.isBackedUp = map['isBackedUp'] ?? false;
+  }
 
-  WalletSchema({this.address, this.type, this.name, this.balance = 0, this.balanceEth = 0, this.isBackedUp = false});
+  Map<String, dynamic> toMap() {
+    return {
+      'address': this.address,
+      'type': this.type,
+      'name': this.name,
+      'balance': this.balance,
+      'balanceEth': this.balanceEth,
+      'isBackedUp': this.isBackedUp,
+    };
+  }
 
   @override
   List<Object> get props => [address, type, name];
 
   @override
   String toString() => 'WalletSchema { address: $address }';
-
-  Future<String> _showDialog(String reason) {
-    return BottomDialog.of(Global.appContext).showInputPasswordDialog(title: NL10ns.of(Global.appContext).verify_wallet_password);
-  }
-
-  Future<String> getPassword({bool showDialogIfCanceledBiometrics = true, bool forceShowInputDialog = false}) async {
-    LOG(tag, usePrint: false).d('getPassword');
-
-    if (forceShowInputDialog) {
-      return _showDialog('force');
-    }
-    final _localAuth = await LocalAuthenticationService.instance;
-    if (_localAuth.isProtectionEnabled) {
-      String password = '';
-      if (password == null) {
-        return _showDialog('no password');
-      } else {
-        bool auth = await _localAuth.authenticate();
-        if (auth) {
-          password = await _secureStorage.get('${SecureStorage.PASSWORDS_KEY}:$address');
-          return password;
-        } else if (showDialogIfCanceledBiometrics) {
-          return _showDialog('auth failed');
-        } else {
-          return null;
-        }
-      }
-    } else {
-      return _showDialog('disabled');
-    }
-  }
-
-  Future<String> getKeystore() async {
-    return await _secureStorage.get('${SecureStorage.NKN_KEYSTORES_KEY}:$address');
-  }
-
-  Future exportWallet(password) async {
-    String keystore = await getKeystore();
-    var wallet = await NknWalletPlugin.openWallet(keystore, password);
-    await _secureStorage.set('${SecureStorage.PASSWORDS_KEY}:$address', password);
-    return wallet;
-  }
-
-  static getWallet({int index = 0}) async {
-    try {
-      SharedPreferences prefs = await SharedPreferences.getInstance();
-      String s = prefs.get('${LocalStorage.NKN_WALLET_KEY}:$index');
-      if (s == null) return null;
-      var walletData = jsonDecode(s);
-      var wallet = WalletSchema(name: walletData['name'], address: walletData['address']);
-      return wallet;
-    } catch (e) {
-      return null;
-    }
-  }
-
-  Future<bool> isDefaultWallet() async {
-    WalletSchema wallet = await getWallet();
-    if (wallet == null || address == null) return false;
-    return wallet.address == address;
-  }
 }
