@@ -1,13 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:get_it/get_it.dart';
-import 'package:nmobile/components/button.dart';
+import 'package:nmobile/components/button/button.dart';
 import 'package:nmobile/components/label.dart';
-import 'package:nmobile/consts/theme.dart';
-import 'package:nmobile/l10n/localization_intl.dart';
+import 'package:nmobile/generated/l10n.dart';
 import 'package:nmobile/schemas/wallet.dart';
-import 'package:nmobile/screens/wallet/create_nkn_wallet.dart';
-import 'package:nmobile/screens/wallet/import_nkn_eth_wallet.dart';
-import 'package:nmobile/utils/extensions.dart';
+import 'package:nmobile/theme/theme.dart';
+
+import 'create_nkn_wallet.dart';
+import 'import_nkn_eth_wallet.dart';
 
 class NoWalletScreen extends StatefulWidget {
   static const String routeName = '/no_wallet';
@@ -17,96 +16,87 @@ class NoWalletScreen extends StatefulWidget {
 }
 
 class _NoWalletScreenState extends State<NoWalletScreen> {
-  final GetIt locator = GetIt.instance;
-
   @override
   Widget build(BuildContext context) {
+    S _localizations = S.of(context);
+
     final screenSize = MediaQuery.of(context);
     return ConstrainedBox(
       constraints: BoxConstraints.expand(),
-      child: Stack(
-        alignment: Alignment.center,
-        children: <Widget>[
-          Container(
-            color: DefaultTheme.backgroundColor4,
-            constraints: BoxConstraints.expand(),
-            child: Scrollbar(
-              child: SingleChildScrollView(
-                child: Container(
-                  height: screenSize.size.height - screenSize.padding.top - screenSize.padding.bottom,
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                    children: [
-                      Expanded(
-                        flex: 0,
-                        child: Padding(
-                          padding: EdgeInsets.only(),
-                          child: Image(image: AssetImage('assets/wallet/pig.png'), width: 120),
-                        ),
+      child: Container(
+        color: DefaultTheme.backgroundColor4,
+        constraints: BoxConstraints.expand(),
+        child: Scrollbar(
+          child: SingleChildScrollView(
+            child: Container(
+              height: screenSize.size.height - screenSize.padding.top - screenSize.padding.bottom,
+              child: Padding(
+                padding: const EdgeInsets.only(top: 60, bottom: 60),
+                child: Flex(
+                  direction: Axis.vertical,
+                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                  children: [
+                    Expanded(
+                      flex: 2,
+                      child: Padding(
+                        padding: EdgeInsets.only(),
+                        child: Image(image: AssetImage('assets/wallet/pig.png'), width: 120),
                       ),
-                      Expanded(
-                        flex: 0,
-                        child: Column(
-                          children: <Widget>[
-                            Label(
-                              NL10ns.of(context).no_wallet_title,
-                              type: LabelType.h2,
+                    ),
+                    Expanded(
+                      flex: 1,
+                      child: Column(
+                        children: <Widget>[
+                          Label(
+                            _localizations.no_wallet_title,
+                            type: LabelType.h2,
+                            dark: true,
+                          ),
+                          Padding(
+                            padding: EdgeInsets.only(top: 16, left: 24, right: 24),
+                            child: Label(
+                              _localizations.no_wallet_desc,
+                              type: LabelType.h4,
                               dark: true,
+                              softWrap: true,
+                              textAlign: TextAlign.center,
                             ),
-                            Padding(
-                              padding: EdgeInsets.only(top: 16, left: 24, right: 24),
-                              child: Label(
-                                NL10ns.of(context).no_wallet_desc,
-                                type: LabelType.h4,
-                                dark: true,
-                                softWrap: true,
-                                textAlign: TextAlign.center,
-                              ),
-                            )
-                          ],
-                        ),
+                          )
+                        ],
                       ),
-                      Expanded(
-                        flex: 0,
+                    ),
+                    Expanded(
+                      flex: 1,
+                      child: Padding(
+                        padding: const EdgeInsets.only(left: 20, right: 20),
                         child: Column(
                           children: <Widget>[
                             Button(
-                              text: NL10ns.of(context).no_wallet_create,
+                              text: _localizations.no_wallet_create,
                               onPressed: () {
-//                                  locator<NavigateService>().pushNamed(CreateNknWalletScreen.routeName);
                                 Navigator.pushNamed(context, CreateNknWalletScreen.routeName);
                               },
                             ),
-                            Button(
-                              text: NL10ns.of(context).no_wallet_import,
-                              backgroundColor: Color(0xFF232D50),
-                              onPressed: () {
-//                                  locator<NavigateService>().pushNamed(ImportNknWalletScreen.routeName);
-                                Navigator.pushNamed(context, ImportWalletScreen.routeName, arguments: WalletType.nkn);
-                              },
-                            ).pad(t: 12),
+                            Padding(
+                              padding: const EdgeInsets.only(top: 12),
+                              child: Button(
+                                text: _localizations.no_wallet_import,
+                                backgroundColor: Color(0xFF232D50),
+                                onPressed: () {
+                                  Navigator.pushNamed(context, ImportWalletScreen.routeName, arguments: {'type': WalletType.nkn});
+                                },
+                              ),
+                            ),
                           ],
-                        ).pad(l: 20, r: 20),
+                        ),
                       ),
-                    ],
-                  ).pad(t: 60, b: 60),
+                    ),
+                  ],
                 ),
               ),
             ),
           ),
-          /*Positioned(
-            top: 0,
-            child: Container(
-              child: Opacity(
-                opacity: 0.25,
-                child: Image(
-                  image: AssetImage('assets/header-background.png'),
-                  fit: BoxFit.cover,
-                ),
-              ),
-            ),
-          ),*/
-        ],
+        ),
       ),
     );
   }
