@@ -1,37 +1,24 @@
 import 'package:flutter/services.dart';
 import 'package:local_auth/local_auth.dart';
-import 'package:nmobile/blocs/account_depends_bloc.dart';
-import 'package:nmobile/helpers/global.dart';
-import 'package:nmobile/helpers/local_notification.dart';
-import 'package:nmobile/helpers/local_storage.dart';
-import 'package:nmobile/utils/log_tag.dart';
-import 'package:synchronized/synchronized.dart';
-import 'package:nmobile/l10n/localization_intl.dart';
+import 'package:nmobile/common/global.dart';
+import 'package:nmobile/generated/l10n.dart';
+import 'package:nmobile/utils/logger.dart';
 
-class LocalAuthenticationService with Tag,AccountDependsBloc {
-  // ignore: non_constant_identifier_names
-  LOG _LOG;
+class LocalAuthenticationService {
 
-  LocalAuthenticationService._() {
-    _LOG = LOG(tag);
-  }
-
-  static LocalAuthenticationService _instance;
-  static Lock _lock = Lock();
-
-  static Future<LocalAuthenticationService> get instance async {
-    if (_instance == null)
-      await _lock.synchronized(() async {
-        if (_instance == null) {
-          final ins = LocalAuthenticationService._();
-          final localStorage = LocalStorage();
-          ins.isProtectionEnabled = (await localStorage.get('${LocalStorage.SETTINGS_KEY}:${LocalStorage.AUTH_KEY}')) as bool ?? false;
-          ins.authType = await ins.getAuthType();
-          _instance = ins;
-        }
-      });
-    return _instance;
-  }
+  // static Future<LocalAuthenticationService> get instance async {
+  //   if (_instance == null)
+  //     await _lock.synchronized(() async {
+  //       if (_instance == null) {
+  //         final ins = LocalAuthenticationService._();
+  //         final localStorage = LocalStorage();
+  //         ins.isProtectionEnabled = (await localStorage.get('${LocalStorage.SETTINGS_KEY}:${LocalStorage.AUTH_KEY}')) as bool ?? false;
+  //         ins.authType = await ins.getAuthType();
+  //         _instance = ins;
+  //       }
+  //     });
+  //   return _instance;
+  // }
 
   final _localAuth = LocalAuthentication();
   bool isProtectionEnabled = false;
@@ -46,14 +33,11 @@ class LocalAuthenticationService with Tag,AccountDependsBloc {
         return BiometricType.fingerprint;
       }
     } on PlatformException catch (e) {
-      _LOG.e('getAuthType | PlatformException', e);
-      LocalNotification.debugNotification('<[DEBUG]> getAuthType', 'PlatformException | ' + e?.message);
+      logger.e(e);
     } on MissingPluginException catch (e) {
-      _LOG.e('getAuthType | MissingPluginException', e);
-      LocalNotification.debugNotification('<[DEBUG]> getAuthType', 'MissingPluginException | ' + e?.message);
+      logger.e(e);
     } catch (e) {
-      _LOG.e('getAuthType | ?', e);
-      LocalNotification.debugNotification('<[DEBUG]> getAuthType', '? | ' + e?.message);
+      logger.e(e);
     }
     return null;
   }
@@ -62,21 +46,18 @@ class LocalAuthenticationService with Tag,AccountDependsBloc {
     if (isProtectionEnabled) {
       try {
         final success = await _localAuth.authenticateWithBiometrics(
-          localizedReason: NL10ns.of(Global.appContext).authenticate_to_access,
+          localizedReason: S.of(Global.appContext).authenticate_to_access,
           useErrorDialogs: false,
           stickyAuth: true,
         );
-        if (!success) LocalNotification.debugNotification('<[DEBUG]> authenticate', 'isProtectionEnabled：$isProtectionEnabled, success: $success');
+
         return success;
       } on PlatformException catch (e) {
-        _LOG.e('authenticate | PlatformException', e);
-        LocalNotification.debugNotification('<[DEBUG]> authenticate', 'PlatformException | ' + e?.message);
+        logger.e(e);
       } on MissingPluginException catch (e) {
-        _LOG.e('authenticate | MissingPluginException', e);
-        LocalNotification.debugNotification('<[DEBUG]> authenticate', 'MissingPluginException | ' + e?.message);
+        logger.e(e);
       } catch (e) {
-        _LOG.e('authenticate | ?', e);
-        LocalNotification.debugNotification('<[DEBUG]> authenticate', '? | ' + e?.message);
+        logger.e(e);
       }
     }
     return false;
@@ -90,10 +71,9 @@ class LocalAuthenticationService with Tag,AccountDependsBloc {
   }
 
   Future<bool> hasBiometrics() async {
-    LocalAuthentication localAuth = new LocalAuthentication();
-    bool canCheck = await localAuth.canCheckBiometrics;
+    bool canCheck = await _localAuth.canCheckBiometrics;
     if (canCheck) {
-      List<BiometricType> availableBiometrics = await localAuth.getAvailableBiometrics();
+      List<BiometricType> availableBiometrics = await _localAuth.getAvailableBiometrics();
 
       if (availableBiometrics.contains(BiometricType.face)) {
         return true;
