@@ -2,34 +2,35 @@ import 'dart:async';
 
 import 'package:dio/dio.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:nkn_sdk_flutter/wallet.dart';
 import 'package:nmobile/blocs/wallet/wallets_bloc.dart';
 import 'package:nmobile/blocs/wallet/wallets_event.dart';
 import 'package:nmobile/blocs/wallet/wallets_state.dart';
-import 'package:nmobile/helpers/global.dart';
-import 'package:nmobile/model/eth_erc20_token.dart';
-import 'package:nmobile/plugins/nkn_wallet.dart';
+import 'package:nmobile/common/global.dart';
+import 'package:nmobile/helpers/nkn_erc20.dart';
 import 'package:nmobile/schemas/wallet.dart';
-import 'package:nmobile/utils/log_tag.dart';
 
-class TaskService with Tag {
+class TaskService {
   Dio dio = Dio();
   WalletsBloc _walletsBloc;
   EthErc20Client _erc20client;
   Timer _queryNknWalletBalanceTask;
   bool _isInit = false;
-  LOG _LOG;
 
-  init() {
+  install() {
     if (!_isInit) {
-      _LOG = LOG(tag);
       _walletsBloc = BlocProvider.of<WalletsBloc>(Global.appContext);
       _erc20client = EthErc20Client();
       _queryNknWalletBalanceTask = Timer.periodic(Duration(seconds: 60), (timer) {
         queryNknWalletBalanceTask();
       });
-
+      queryNknWalletBalanceTask();
       _isInit = true;
     }
+  }
+
+  uninstall (){
+    _queryNknWalletBalanceTask?.cancel();
   }
 
   queryNknWalletBalanceTask() {
@@ -39,19 +40,17 @@ class TaskService with Tag {
       state.wallets.forEach((w) {
         if (w.type == WalletSchema.ETH_WALLET) {
           futures.add(_erc20client.getBalance(address: w.address).then((balance) {
-            _LOG.w('${w.name} | balance: ${balance.ether}');
             w.balanceEth = balance.ether;
             _walletsBloc.add(UpdateWallet(w));
           }));
           futures.add(_erc20client.getNknBalance(address: w.address).then((balance) {
             if (balance != null) {
-              _LOG.w('${w.name} | Nkn balance: ${balance.ether}');
               w.balance = balance.ether;
               _walletsBloc.add(UpdateWallet(w));
             }
           }));
-        } else {
-          futures.add(NknWalletPlugin.getBalanceAsync(w.address).then((balance) {
+        } else if(w.type == WalletSchema.NKN_WALLET) {
+          futures.add(Wallet.getBalanceByAddr(w.address).then((balance) {
             w.balance = balance;
             _walletsBloc.add(UpdateWallet(w));
           }));
@@ -63,15 +62,4 @@ class TaskService with Tag {
     }
   }
 
-//  queryTopicCountTask() async {
-//    if (Global.currentChatDb == null) {
-//      return;
-//    }
-//    var topics = await TopicSchema.getAllTopic();
-//    if (topics != null) {
-//      topics.forEach((x) {
-//        x.getTopicCount();
-//      });
-//    }
-//  }
 }
